requirements.txt
feedparser
newspaper3k
lxml
lxml_html_clean
nltk
rich
